+++ tests/try_trait.rs
@@ -1,5 +1,5 @@
-#![cfg(feature = "try_trait")]
-#![feature(try_trait_v2)]
+#![cfg(all(feature = "try_trait", try_trait_nightly))]
+#![cfg_attr(all(feature = "try_trait", try_trait_nightly), feature(try_trait_v2))]
 
 use stillwater::Validation;